refactor: Update import paths for RegistrationLayout in authentication
