+++ src/app/(authentication)/layout.tsx
@@ -1,4 +1,4 @@
-import RegistrationLayout from "./RegistrationLayout";
+import RegistrationLayout from "@/components/layouts/RegistrationLayout";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return <RegistrationLayout>{children}</RegistrationLayout>;
