+++ src/app/(team)/layout.tsx
@@ -1,4 +1,4 @@
-import RegistrationLayout from "@/app/(authentication)/RegistrationLayout";
+import RegistrationLayout from "@/components/layouts/RegistrationLayout";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
