+++ src/components/layouts/RegistrationLayout.tsx
